style: drop redundant inline, typename, parentheses and casts
Mechanical, applied by clang-tidy across 77 sites in 28 files: `inline` on
functions that are already implicitly inline, `typename` where C++20 no longer
requires it, redundant parentheses, and casts to a type the expression already
has.

readability-redundant-member-init was attempted alongside these and is left
disabled: its fix removes the `{}` from `T member {};`, which is not redundant
in the aggregates this codebase initializes with designated initializers that
omit fields -- the member is then left uninitialized and the build fails on
-Wmissing-designated-field-initializers. The rationale is recorded in the config.

+++ src/vtparser/Parser-impl.h
@@ -501,7 +501,7 @@ void Parser<EventListener, TraceStateChanges>::processOnceViaStateMachine(uint8_
 
     auto const s = static_cast<size_t>(_state);
 
-    if (auto const t = Table.transitions[s][static_cast<uint8_t>(ch)]; t != State::Undefined)
+    if (auto const t = Table.transitions[s][ch]; t != State::Undefined)
     {
         // std::cout << std::format("VTParser: Transitioning from {} to {}", _state, t);
         handle(ActionClass::Leave, Table.exitEvents[s], ch);
